+++ packages/ripple-binary-codec/HISTORY.md
@@ -4,6 +4,17 @@
 ### Added
 - Support for the XChainBridge amendment.
 
+## 1.8.0 (2023-08-07)
+* Added Clawback transaction type
+
+## 1.7.1 (2023-07-18)
+### Fixed
+* Passing Definitions from `STObject` to `STArray` back to `STObject`, both for signing & decoding (to JSON)
+
+## 1.7.0 (2023-07-12)
+### Added
+* `NetworkId` field support
+
 ## 1.6.0 (2023-06-13)
 ### Added
 - Allow custom type definitions to be used for encoding/decoding transactions at runtime (e.g. for sidechains/new amendments)